Add Task1_1

+++ Task1/Task1_1/Program.cs
@@ -7,16 +7,16 @@ namespace Task1_1
         static void Main(string[] args)
         {
             Console.WriteLine("Введите первую сторону прямоугольника: ");
-            double a = double.Parse(Console.ReadLine());
+            int a = int.Parse(Console.ReadLine());
             Console.WriteLine("Введите вторую сторону прямоугольника: ");
-            double b = double.Parse(Console.ReadLine());
+            int b = int.Parse(Console.ReadLine());
             if (a <= 0 || b <= 0)
             {
                 Console.WriteLine("Вы ввели 0 или отрицательное число");
             }
             else
             {
-                double result = a * b;
+                int result = a * b;
                 Console.WriteLine("Площадь прямоугольника: " + result);
             }
